changed makeANote
makeANote function now creates an object so now an object gets pushed to listOfNotes instead of just a string. This object has more properties of the note than just a name and octave... although I could have just used a longer string with a system for encoding info about a note into the string. eg f4sf could mean f4 sharp but with displaySharp=false (because it is in the key sig and i don't want a redundant sharp sign in this case) maybe I will implement this so i can further procrastinate mastery of javascript objects and obstinately do more with less.

renamed musicnote.html to index.html
created new main.js so now index.html has zero javascript in it. and zero css.